[CL-1576] Extract setup logic into seperate file
